Added the ability to load levels randomly
There is now a LevelData.plist file which contains the level
name, difficulty and the number of helmet and umbrella uses
allowed. This is loaded when the game is first run.

When the player starts a new game, a random level is chosen
with the chosen difficulty.

Added a Level class to contain this data, which also has an
extra spawnPoint variable which (set from the entrance object's
coords) so the lemmings spawn wherever the entrance is placed.