Lots more added to the AI
Added extra trigger objects throughout the levels which add as
'decision nodes' (as well as trapdoors). Mainly used for the end
of platforms, but may extend for use anywhere.

At the moment, the agent acts randomly at each 'decision object',
which is simulating the 'learning mode'. Have also added the option
to randomise the chosen action (var in Constants - the larger the
size, the smaller the probability of a random action to be chosen).

Added relevant code to calculate which options an agent has available,
creating an array of options.

Also tidied up the collision detection to work with the decision nodes.

Added a method to the Utils class which gets the string of any passed
decision enum.

Simplified the game rating algorithm which now works a bit better.